requirements.txt
django
djangorestframework
djangorestframework-jwt
django-cors-headers
pillow
gunicorn
django-heroku
